Solve problem with stack not showing
This was caused by setting the upper limit of the for loop
to a variable that was initialized at zero.
This cause the stack to be empty on program execution.
To solve this I had to set the upper limit to a new variable
with the number of elements of the array